+++ config/lineConfig.js
@@ -1,0 +1,6 @@
+const config = {
+  channelAccessToken: process.env.CHANNEL_ACCESS_TOKEN,
+  channelSecret: process.env.CHANNEL_SECRET,
+};
+
+module.exports = config;
